requirements.txt
alpha_vantage==2.3.1
pandas==2.2.0
numpy==1.26.3
pymongo==4.6.1
newsapi-python==0.2.7
kafka-python==2.0.2
tweepy==4.14.0
praw==7.7.1
requests==2.32.3
fredapi==0.5.1
beautifulsoup4==4.12.3
textblob==0.17.1
fastavro==1.9.4
